Refactored model.py to only contain BaseModel and moved MedicareModel to medicare_model.py
